created file searchHotels
