+++ index.html
@@ -712,14 +712,5 @@
       </div>
     </footer>
     <script type="module" src="src/js/main.js"></script>
-    <script type="module" src="src/js/sections/createdSectionAvailable.js"></script>
-<!--    <script-->
-<!--      type="module"-->
-<!--      src="src/js/desktopFormFilter/searchHotelsFirstInput.js"-->
-<!--    ></script>-->
-<!--    <script-->
-<!--      type="module"-->
-<!--      src="src/js/desktopFormFilter/searchHotelsThreeInput.js"-->
-<!--    ></script>-->
   </body>
 </html>
