+++ src/js/constants-utils/constants.js
@@ -1,8 +1,10 @@
-
-
-
 const destinationsEl = document.getElementById('destinations'); // получил секцию destinations по id
 const mainEl = document.getElementById('main'); // получил tag main по id
-const URL = 'https://if-student-api.onrender.com/api/hotels/popular';
+const urlPopularHotel = 'https://if-student-api.onrender.com/api/hotels/popular';
+const urlAllHotels = new URL('https://if-student-api.onrender.com/api/hotels');
+const { desktopForm } = document.forms;
 
-export { destinationsEl, mainEl, URL }
+const offerEl = document.getElementById('offer'); // получил секцию offer по id
+
+
+export { destinationsEl, mainEl, urlPopularHotel, offerEl, desktopForm, urlAllHotels };
