--- src/js/desktopFormFilter/searchHotelsThreeInput.js
@@ -1,5 +1,0 @@
-const inputEl = document.getElementById('desktopForm');
-
-const formdate = new FormData(inputEl);
-
-console.log(formdate);
